model/requirements.txt
albumentations
absl-py==1.0.0
astunparse==1.6.3
cachetools==4.2.4
catboost==1.0.3
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.3
cycler==0.11.0
Cython==0.29.24
datatable==1.0.0
Deprecated==1.2.13
fonttools==4.28.2
gast==0.3.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.19
grpcio==1.42.0
h5py==2.10.0
idna==3.3
imageio==2.13.1
importlib-metadata==4.8.2
itsdangerous==2.0.1
jieba==0.42.1
Jinja2==3.0.3
joblib==1.1.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lightgbm==3.3.1
llvmlite==0.37.0
Markdown==3.3.6
MarkupSafe==2.0.1
matlabengineforpython===R2021a
matplotlib==3.5.0
networkx==2.6.3
nltk==3.6.5
numba==0.54.1
numexpr==2.8.0
numpy==1.18.5
oauthlib==3.1.1
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
plotly==5.4.0
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
redis==4.0.2
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.4.1
setuptools-scm==6.3.2
six==1.16.0
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
tomli==1.2.2
torch==1.7.1+cu101
torchaudio==0.7.2
torchvision==0.8.2+cu101
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
xgboost==1.5.1
zipp==3.6.0